fix: unblock CI — repair pre-existing build/vet/test breaks on master
master currently fails go build ./..., go vet ./..., and go test ./...
independent of any feature work: a stale unused import, a dropped
CreateSession argument, missing imports and float64/int64 mismatches
in deposit_handler.go, two router-referenced UserHandler methods that
were never implemented, a missing config field, and several tests
whose mocks drifted behind real service logic (missing ListByCircle/
IsMember/GetAnnouncementByID/FindByID expectations, stale float64
assertions against now-int64 money fields).

Fixes only what's needed to get build/vet/test green again; no
behavior changes beyond adding 501 stubs for the two dangling KYC
routes so the package compiles.

+++ internal/api/handler/deposit_persistence_test.go
@@ -117,7 +117,7 @@ func TestDepositHandler_InitiateWithdraw_PersistsRecord(t *testing.T) {
 	}
 	require.NotNil(t, wd)
 	assert.Equal(t, "test-user-123", wd.UserID)
-	assert.Equal(t, float64(100), wd.AmountUSDC)
+	assert.Equal(t, int64(100), wd.AmountUSDC)
 	require.NotNil(t, wd.YellowCardTxID)
 	assert.Equal(t, "s-1", *wd.YellowCardTxID)
 }
